Moved tests to proper folder
